+++ src/Extension.php
@@ -26,7 +26,7 @@ try {
     ));
 }
 
-if ($phpUnitVersionSeries->major()->isOneOf(Version\Major::fromInt(8), Version\Major::fromInt(9))) {
+if ($phpUnitVersionSeries->major()->isOneOf(Version\Major::fromInt(7), Version\Major::fromInt(8), Version\Major::fromInt(9))) {
     /**
      * @internal
      */